DWL 3.0009c
- Afegida una nova funció privada a DApp que em deixa capturar els Access Violations amb un catch... (això es per que tinc un bug raro.. que crec que ve del DAnimacion...
- Modificat el projecte per utilitzxar la funció /SHa del compilador, per detectar Acces Violations i demes errors chungus en un catch.

+++ DWL3/DConsola.h
@@ -21,7 +21,7 @@
 
 			const BOOL              MostrarUltimoError();
 		protected:
-			DWORD				   _UltimoTick;
+			ULONGLONG			   _UltimoTick;
 			HANDLE                 _Consola;
 			HANDLE                 _Mutex;
 		};
